+++ docs/source/conf.py
@@ -13,7 +13,7 @@ sys.path.append(os.path.abspath('../../src/'))
 project = 'atompy'
 copyright = 'Max Kircher, CC BY-NC 4.0'
 author = 'Max Kircher'
-release = '2.1.0'
+release = '3.0.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -22,7 +22,7 @@ extensions = [
     "sphinx.ext.autodoc",
     "numpydoc",
     "sphinx.ext.viewcode",
-    "sphinx_rtd_theme",
+    # "sphinx_rtd_theme",
     'matplotlib.sphinxext.plot_directive',
 ]
 autodoc_type_aliases = {
@@ -37,8 +37,12 @@ exclude_patterns = []
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'sphinx_rtd_theme'
+html_theme = 'pydata_sphinx_theme'
 html_static_path = ['_static']
 
 html_theme_options = {
+    "show_toc_level": 4,
+    "collapse_navigation": False,
+    "navigation_with_keys": False,
+    "show_nav_level": 4,
 }