+++ src/atompy/physics/__init__.py
@@ -1,0 +1,1 @@
+from atompy.physics._physics import *